@@ -4,29 +4,34 @@
     "keywords": ["compatibility", "database", "reference"],
     "type": "library",
     "license": "BSD-3-Clause",
-    "homepage": "http://php5.laurent-laville.org/compatinfo/",
+    "homepage": "http://bartlett.laurent-laville.org/php-compatinfo/",
     "support": {
         "source": "https://github.com/llaville/php-compatinfo-db",
         "issues": "https://github.com/llaville/php-compatinfo-db/issues"
     },
     "require": {
-        "php": "^7.1|^8.0",
-        "composer/semver": "^1.0|^2.0|^3.0",
-        "ext-curl": "*",
-        "ext-intl": "*",
-        "ext-libxml": "*",
-        "ext-openssl": "*",
-        "ext-pcre": "*",
-        "ext-spl": "*",
+        "php": "^7.2|^8.0",
         "ext-json": "*",
-        "ext-pdo_sqlite": "*",
-        "symfony/console": "^4.4|^5.0",
+        "ext-pcre": "*",
+        "ext-pdo": "*",
+        "ext-phar": "*",
+        "ext-spl": "*",
         "composer/package-versions-deprecated": "^1.8",
-        "league/tactician": "^1.0",
-        "laminas/laminas-diagnostics": "^1.3"
+        "composer/semver": "^1.0|^2.0|^3.0",
+        "doctrine/orm": "^2.7",
+        "symfony/config": "^4.4|^5.0",
+        "symfony/console": "^4.4|^5.0",
+        "symfony/dependency-injection": "^4.4|^5.0",
+        "symfony/event-dispatcher": "^4.4|^5.0",
+        "symfony/messenger": "^4.4|^5.0",
+        "symfony/requirements-checker": "^2.0",
+        "symfony/stopwatch": "^4.4|^5.0"
     },
     "require-dev": {
-        "phpunit/php-timer": "^2.0"
+        "composer/composer": "^2.0",
+        "humbug/box": "^3.9",
+        "psr/log": "^1.0",
+        "symfony/phpunit-bridge": "^5.1"
     },
     "authors": [
         {
@@ -48,7 +53,27 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Bartlett\\Tests\\CompatInfoDb\\": "tests/"
+            "Bartlett\\CompatInfoDb\\Tests\\": "tests/"
         }
+    },
+    "scripts": {
+        "compile-box": "box compile",
+        "post-install-cmd": [
+            "Bartlett\\CompatInfoDb\\ComposerScripts::preparePharMetadata",
+            "@compile-box"
+        ],
+        "post-update-cmd": [
+            "Bartlett\\CompatInfoDb\\ComposerScripts::preparePharMetadata",
+            "@compile-box"
+        ]
+    },
+    "scripts-descriptions": {
+        "compile-box": "Compiles application with box-project/box to build PHAR version"
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true,
+    "config": {
+        "sort-packages": true,
+        "optimize-autoloader": true
     }
 }
