@@ -1,6 +1,7 @@
-<?php
+<?php declare(strict_types=1);
+
 /**
- * Unit tests for PHP_CompatInfo, Generic extension base class.
+ * Unit tests for PHP_CompatInfo_Db, Generic extension base class.
  *
  * PHP version 7
  *
@@ -10,135 +11,89 @@
  * @author     Laurent Laville <pear@laurent-laville.org>
  * @author     Remi Collet <Remi@FamilleCollet.com>
  * @license    https://opensource.org/licenses/BSD-3-Clause The 3-Clause BSD License
- * @link       http://php5.laurent-laville.org/compatinfo/
+ * @link       http://bartlett.laurent-laville.org/php-compatinfo/
  */
 
-namespace Bartlett\Tests\CompatInfoDb\Reference;
+namespace Bartlett\CompatInfoDb\Tests\Reference;
 
-use Bartlett\CompatInfoDb\ExtensionFactory;
+use Bartlett\CompatInfoDb\Domain\Factory\ExtensionFactory;
+use Bartlett\CompatInfoDb\Domain\Factory\ExtensionVersionProviderInterface;
+use Bartlett\CompatInfoDb\Domain\Factory\ExtensionVersionProviderTrait;
+use Bartlett\CompatInfoDb\Domain\Factory\LibraryVersionProviderTrait;
+use Bartlett\CompatInfoDb\Domain\ValueObject\Dependency;
+use Bartlett\CompatInfoDb\Domain\ValueObject\Extension;
 
 use Composer\Semver\Semver;
+
 use PHPUnit\Framework\ExpectationFailedException;
 
+use Generator;
+use ReflectionClass;
+use ReflectionException;
+use ReflectionExtension;
+use ReflectionFunction;
+use ReflectionMethod;
+use function array_filter;
+use function array_key_exists;
+use function array_keys;
+use function array_map;
+use function array_pop;
+use function array_push;
+use function array_unique;
+use function class_exists;
+use function dechex;
+use function defined;
+use function dirname;
+use function end;
+use function explode;
+use function extension_loaded;
+use function get_called_class;
+use function get_defined_constants;
+use function get_extension_funcs;
+use function hexdec;
+use function implode;
+use function in_array;
+use function ini_get;
+use function interface_exists;
+use function is_array;
+use function phpversion;
+use function sprintf;
+use function str_replace;
+use function str_split;
+use function strcasecmp;
+use function strlen;
+use function strtolower;
+use function version_compare;
+use const DIRECTORY_SEPARATOR;
+
 /**
- * Tests for the PHP_CompatInfo, retrieving components informations
- * about any extension.
- *
- * @category   PHP
- * @package    PHP_CompatInfo_Db
- * @subpackage Tests
- * @author     Laurent Laville <pear@laurent-laville.org>
- * @author     Remi Collet <Remi@FamilleCollet.com>
- * @license    https://opensource.org/licenses/BSD-3-Clause The 3-Clause BSD License
- * @link       http://php5.laurent-laville.org/compatinfo/
- * @since      Class available since Release 3.0.0RC1 of PHP_CompatInfo
- * @since      Class available since Release 1.0.0alpha1 of PHP_CompatInfo_Db
+ * @since Release 3.0.0RC1 of PHP_CompatInfo
+ * @since Release 1.0.0alpha1 of PHP_CompatInfo_Db
  */
-abstract class GenericTest extends \PHPUnit\Framework\TestCase
+abstract class GenericTest extends TestCase implements ExtensionVersionProviderInterface
 {
-    const REF_ELEMENT_INI       = 1;
-    const REF_ELEMENT_CONSTANT  = 2;
-    const REF_ELEMENT_FUNCTION  = 3;
-    const REF_ELEMENT_INTERFACE = 4;
-    const REF_ELEMENT_CLASS     = 5;
-    const REF_ELEMENT_METHOD    = 6;
-    const REF_ELEMENT_CONST     = 7;  // class constant
+    use LibraryVersionProviderTrait;
+    use ExtensionVersionProviderTrait;
 
     protected static $obj = null;
-    protected static $ref = null;
-    protected static $ext = null;
-
-    protected $extension;
 
     // Could be defined in Reference but missing (system dependant)
-    protected static $optionalreleases    = array();
-    protected static $optionalcfgs        = array();
-    protected static $optionalconstants   = array();
-    protected static $optionalfunctions   = array();
-    protected static $optionalclasses     = array();
-    protected static $optionalinterfaces  = array();
-    protected static $optionalmethods     = array();
+    protected static $optionalreleases    = [];
+    protected static $optionalcfgs        = [];
+    protected static $optionalconstants   = [];
+    protected static $optionalfunctions   = [];
+    protected static $optionalclasses     = [];
+    protected static $optionalinterfaces  = [];
+    protected static $optionalmethods     = [];
 
     // Could be present but missing in Reference (alias, ...)
-    protected static $ignoredcfgs          = array();
-    protected static $ignoredconstants     = array();
-    protected static $ignoredfunctions     = array();
-    protected static $ignoredclasses       = array();
-    protected static $ignoredinterfaces    = array();
-    protected static $ignoredmethods       = array();
-    protected static $ignoredconsts        = array();
-
-    // References fully documented
-    protected static $extensions =  array(
-        'amqp',
-        'date',
-        'exif',
-        'gender',
-        'geoip',
-        'gmp',
-        'haru',
-        'htscanner',
-        'imagick',
-        'igbinary',
-        'inclued',
-        'intl',
-        'jsmin',
-        'ldap',
-        'lzf',
-        'mailparse',
-        'mongo',
-        'msgpack',
-        'mysqli',
-        'oauth',
-        'openssl',
-        'pdflib',
-        'pgsql',
-        'pthreads',
-        'rar',
-        'redis',
-        'reflection',
-        'riak',
-        'soap',
-        'sockets',
-        'solr',
-        'sphinx',
-        'spl',
-        'sqlite3',
-        'ssh2',
-        'stomp',
-        'sync',
-        'tidy',
-        'uopz',
-        'uploadprogress',
-        'varnish',
-        'xdebug',
-        'xmldiff',
-        'xmlrpc',
-        'xsl',
-        'Zend OPcache',
-        'zip',
-    );
-
-    /**
-     * {@inheritDoc}
-     */
-    public function __construct($name = null, array $data = [], $dataName = '')
-    {
-        parent::__construct($name, $data, $dataName);
-
-        $parts = explode('\\', get_class($this));
-
-        self::$ext = $name = strtolower(
-            str_replace('ExtensionTest', '', end($parts))
-        );
-
-        // special case(s)
-        if ('zendopcache' === $name) {
-            self::$ext = $name = 'zend opcache';
-        }
-
-        self::$ext = $this->extension = $name;
-    }
+    protected static $ignoredcfgs          = [];
+    protected static $ignoredconstants     = [];
+    protected static $ignoredfunctions     = [];
+    protected static $ignoredclasses       = [];
+    protected static $ignoredinterfaces    = [];
+    protected static $ignoredmethods       = [];
+    protected static $ignoredconsts        = [];
 
     /**
      * Sets up the shared fixture.
@@ -148,25 +103,30 @@ abstract class GenericTest extends \PHPUnit\Framework\TestCase
      */
     public static function setUpBeforeClass(): void
     {
-        self::$optionalreleases = array();
+        $container = require implode(DIRECTORY_SEPARATOR, [dirname(__DIR__, 2), 'config', 'container.php']);
 
-        if (!is_object(self::$obj) || self::$obj->getName() !== self::$ext) {
-            self::$obj = new ExtensionFactory(self::$ext);
-        }
+        self::$optionalreleases = [];
 
-        $currentVersion = self::$obj->getCurrentVersion();
+        $parts = explode('\\', get_called_class());
+        $name = strtolower(
+            str_replace('ExtensionTest', '', end($parts))
+        );
 
+        $factory = $container->get(ExtensionFactory::class);
+        self::$obj = $factory->create($name);
+
+        $currentVersion = phpversion($name);
         if ($currentVersion === false) {
             // extension did not provide any version information
             return;
         }
 
-        $releases = array_keys(self::$obj->getReleases());
+        $releases = self::$obj->getReleases();
 
         // platform dependant
-        foreach ($releases as $rel_version) {
-            if (version_compare($currentVersion, $rel_version, 'lt')) {
-                array_push(self::$optionalreleases, $rel_version);
+        foreach ($releases as $release) {
+            if (version_compare($currentVersion, $release->getVersion(), 'lt')) {
+                array_push(self::$optionalreleases, $release->getVersion());
             }
         }
     }
@@ -190,154 +150,58 @@ abstract class GenericTest extends \PHPUnit\Framework\TestCase
         self::$ignoredconsts      = [];
     }
 
+    protected function setUp(): void
+    {
+        $name = self::$obj->getName();
+        // special case(s)
+        if ('opcache' === $name) {
+            $name = 'zend opcache';
+        }
+
+        if (!extension_loaded($name)) {
+            $this->markTestSkipped(
+                sprintf('Extension %s is required.', $name)
+            );
+        }
+    }
+
     /**
      * Generic Reference validator and producer
+     *
+     * @param array $elements
+     * @param string $opt
+     * @return Generator
      */
-    public function provideReferenceValues($methodName)
+    private function provideReferenceValues(array $elements, string $opt): Generator
     {
-        static $obj;
+        $extVersion = $this->getExtensionVersion(self::$obj->getName());
 
-        $this->checkExtension();
+        foreach ($elements as $name => $element) {
+            $range = [
+                'ext.min' => $extVersion,
+                'ext.max' => $element->getExtMax(),
+                'php.min' => $element->getPhpMin(),
+                'php.max' => $element->getPhpMax(),
+            ];
 
-        if (!is_object($obj) || $obj->getName() !== $this->extension) {
-            $obj = new ExtensionFactory($this->extension);
-        }
-
-        if ('testGetIniEntriesFromReference' === $methodName) {
-            $refElementType = self::REF_ELEMENT_INI;
-            $elements = $obj->getIniEntries();
-            $opt = 'optionalcfgs';
-
-        } elseif ('testGetFunctionsFromReference'  === $methodName) {
-            $refElementType = self::REF_ELEMENT_FUNCTION;
-            $elements = $obj->getFunctions();
-            $opt = 'optionalfunctions';
-
-        } elseif ('testGetConstantsFromReference'  === $methodName) {
-            $refElementType = self::REF_ELEMENT_CONSTANT;
-            $elements = $obj->getConstants();
-            $opt = 'optionalconstants';
-
-        } elseif ('testGetClassesFromReference' === $methodName) {
-            $refElementType = self::REF_ELEMENT_CLASS;
-            $elements = $obj->getClasses();
-            $opt = 'optionalclasses';
-
-        } elseif ('testGetInterfacesFromReference' === $methodName) {
-            $refElementType = self::REF_ELEMENT_INTERFACE;
-            $elements = $obj->getInterfaces();
-            $opt = 'optionalinterfaces';
-
-        } elseif ('testGetClassMethodsFromReference' === $methodName) {
-            $refElementType = self::REF_ELEMENT_METHOD;
-            $elements = [];
-
-            $methods = array_merge(
-                $obj->getClassMethods(),
-                $obj->getClassStaticMethods()
-            );
-
-            foreach ($methods as $class => $values) {
-                foreach ($values as $method => $range) {
-                    $elements[$class.'::'.$method] = $range;
-                }
-            }
-            $opt = 'optionalmethods';
-
-        } else {
-            $elements = [];
-        }
-
-        foreach ($elements as $name => $range) {
             if (!empty($range['optional'])) {
                 self::${$opt}[] = $name;
                 continue;
             }
 
-            $libs = array();
-            foreach ($range as $key => $val) {
-                if (strpos($key, 'lib_') === 0) {
-                    if (!empty($val)) {
-                        $libs[$key] = $val;
-                    }
-                }
-            }
-
-            foreach($libs as $lib => $constraint) {
-                $lib = str_replace('lib_', '', $lib);
-                $ver = self::lib($lib, 'version_text');
-
-                if (!Semver::satisfies($ver, $constraint)) {
+            foreach ($element->getDependencies() as $dependency) {
+                /** @var Dependency $dependency */
+                $ver = $this->getPrettyVersion($dependency->getName());
+                if (!Semver::satisfies($ver, $dependency->getConstraint())) {
                     self::${$opt}[] = $name;
                     continue 2;
                 }
             }
-            yield [$name, $range, $refElementType];
+            yield [$name, $range];
         }
     }
 
-    protected static function lib($name, $key = 'version_number')
-    {
-        if ('curl' == $name
-            && function_exists('curl_version')
-        ) {
-            $meta = curl_version();
-            $meta['version_text'] = $meta['version'];
-
-        } elseif ('libxml' == $name) {
-            $meta = array(
-                'version_number' => defined('LIBXML_DOTTED_VERSION')
-                    ? self::toNumber(\LIBXML_DOTTED_VERSION) : false,
-                'version_text'   => defined('LIBXML_DOTTED_VERSION')
-                    ? \LIBXML_DOTTED_VERSION : false,
-            );
-
-        } elseif ('intl' == $name) {
-            $meta = array(
-                'version_number' => defined('INTL_ICU_VERSION')
-                    ? self::toNumber(\INTL_ICU_VERSION) : false,
-                'version_text'   => defined('INTL_ICU_VERSION')
-                    ? \INTL_ICU_VERSION : false,
-            );
-
-        } elseif ('openssl' == $name) {
-            $meta = array(
-                'version_number' => defined('OPENSSL_VERSION_NUMBER')
-                    ? \OPENSSL_VERSION_NUMBER : false,
-                'version_text'   => defined('OPENSSL_VERSION_TEXT')
-                    ? self::toText(\OPENSSL_VERSION_NUMBER) : false,
-            );
-
-        } elseif ('imagemagick' == $name) {
-            if (method_exists('Imagick', 'getVersion')) {
-                $v = \Imagick::getVersion();
-                if (preg_match('/ImageMagick ([0-9]+\.[0-9]+\.[0-9]+)/', $v['versionString'], $matches)) {
-                    $meta = array(
-                        'version_number' => $v['versionNumber'],
-                        'version_text'   => $matches[1],
-                    );
-                }
-            }
-
-        } elseif ('sqlite3' == $name) {
-            if (method_exists('sqlite3', 'version')) {
-                $v = \SQLite3::version();
-                $meta = array(
-                    'version_number' => $v['versionNumber'],
-                    'version_text'   => $v['versionString'],
-                );
-            }
-        }
-        if (isset($meta)) {
-            if (isset($key) && array_key_exists($key, $meta)) {
-                return $meta[$key];
-            }
-            return $meta;
-        }
-        return false;
-    }
-
-    protected static function toText($number)
+    protected static function toText($number): string
     {
         $hex = dechex(($number & ~ 15) / 16);
 
@@ -350,7 +214,7 @@ abstract class GenericTest extends \PHPUnit\Framework\TestCase
         return implode('.', array_map('hexdec', $arr));
     }
 
-    protected static function toNumber($text)
+    protected static function toNumber($text): int
     {
         $arr = explode('.', $text);
         $arr = array_map('dechex', $arr);
@@ -368,41 +232,17 @@ abstract class GenericTest extends \PHPUnit\Framework\TestCase
     }
 
     /**
-     *
+     * @param string $element
+     * @param array $range
+     * @param array $optional
+     * @param array $ignored
+     * @return bool|null NULL if reference should be skipped, boolean otherwise
      */
-    protected function checkValuesFromReference($element, $range, $refElementType)
+    private function checkValuesFromReference(string $element, array $range, array $optional, array $ignored): ?bool
     {
         if (in_array($range['ext.min'], self::$optionalreleases)) {
-            return;
+            return null;
         }
-
-        if (self::REF_ELEMENT_INI == $refElementType) {
-            $optional = self::$optionalcfgs;
-            $ignored = self::$ignoredcfgs;
-        } elseif (self::REF_ELEMENT_CONSTANT == $refElementType) {
-            $optional = self::$optionalconstants;
-            $ignored = self::$ignoredconstants;
-        } elseif (self::REF_ELEMENT_FUNCTION == $refElementType) {
-            $optional = self::$optionalfunctions;
-            $ignored = self::$ignoredfunctions;
-        } elseif (self::REF_ELEMENT_INTERFACE == $refElementType) {
-            $optional = self::$optionalinterfaces;
-            $ignored = self::$ignoredinterfaces;
-        } elseif (self::REF_ELEMENT_CLASS == $refElementType) {
-            $optional = self::$optionalclasses;
-            $ignored = self::$ignoredclasses;
-        } elseif (self::REF_ELEMENT_METHOD == $refElementType) {
-            $optional = self::$optionalmethods;
-            $ignored = self::$ignoredmethods;
-        }
-
-        $min = $range['php.min'];
-        $max = $range['php.max'];
-
-        $EXT_VERSION = self::$obj->getCurrentVersion();
-
-        $emin = $range['ext.min'];
-        $emax = $range['ext.max'];
 
         if (array_key_exists('php.excludes', $range)) {
             if (in_array(PHP_VERSION, $range['php.excludes'])) {
@@ -410,155 +250,72 @@ abstract class GenericTest extends \PHPUnit\Framework\TestCase
                 array_push($optional, $element);
             }
         }
-        if (!in_array($element, $optional)
-            && (empty($min) || version_compare(PHP_VERSION, $min) >= 0)
-            && (empty($max) || version_compare(PHP_VERSION, $max) <= 0)
-            && (!empty($emin) && version_compare($EXT_VERSION, $emin) >= 0)
-            && (!empty($emax) && version_compare($EXT_VERSION, $emax) <= 0)
-        ) {
-            // Should be there except if set as optional
-            $this->assertShouldBeThere($element, $refElementType);
+
+        if (in_array($element, $optional) || in_array($element, $ignored)) {
+            return null;
         }
-        if (!in_array($element, $ignored)) {
-            if (($min && version_compare(PHP_VERSION, $min) < 0)
-                || ($max && version_compare(PHP_VERSION, $max) > 0)
-            ) {
-                // Should not be there except if ignored
-                $this->assertShouldNotBeThere($element, $refElementType, $min, $max);
+
+        $min = $range['php.min'];
+        $max = $range['php.max'];
+
+        $emin = $range['ext.min'];
+        $emax = $range['ext.max'];
+
+        $deprecated = $range['deprecated'] ?? '';
+
+        if (!empty($deprecated)) {
+            $shouldBeThere = version_compare(PHP_VERSION, $deprecated, 'le');
+
+            // used also for elements that were moved from one extension to another;
+            // i.e with `utf8_encode` (from `xml` to `standard` extension)
+
+            if (!$shouldBeThere) {
+                return null; // ignore it !
             }
         }
-    }
 
-    protected function assertShouldBeThere($element, $refElementType)
-    {
-        if (self::REF_ELEMENT_INI == $refElementType) {
-            $this->assertNotSame(
-                ini_get($element),
-                false,
-                "INI '$element', found in Reference, does not exists."
-            );
+        $extVersion = $this->getExtensionVersion(self::$obj->getName());
 
-        } elseif (self::REF_ELEMENT_FUNCTION == $refElementType) {
-            $this->assertTrue(
-                function_exists($element),
-                "Function '$element', found in Reference, does not exists."
-            );
-
-        } elseif (self::REF_ELEMENT_CONSTANT == $refElementType) {
-            $this->assertTrue(
-                defined($element),
-                "Constant '$element', found in Reference, does not exists."
-            );
-
-        } elseif (self::REF_ELEMENT_CLASS == $refElementType) {
-            $this->assertTrue(
-                class_exists($element, false),
-                "Class '$element', found in Reference, does not exists."
-            );
-
-        } elseif (self::REF_ELEMENT_INTERFACE == $refElementType) {
-            $this->assertTrue(
-                interface_exists($element, false),
-                "Interface '$element', found in Reference, does not exists."
-            );
-
-        } elseif (self::REF_ELEMENT_METHOD == $refElementType) {
-            list ($object, $method) = explode('::', $element);
-            $this->assertTrue(
-                method_exists($object, $method),
-                "Class Method '$element', found in Reference, does not exists."
-            );
+        if (!empty($min)) {
+            $shouldBeThere = version_compare(PHP_VERSION, $min, 'ge');
+        } else {
+            $shouldBeThere = false;
         }
-    }
-
-    protected function assertShouldNotBeThere($element, $refElementType, $min, $max)
-    {
-        if (self::REF_ELEMENT_INI == $refElementType) {
-            $this->assertFalse(
-                ini_get($element),
-                "INI '$element', found in Reference ($min,$max), exists."
-            );
-
-        } elseif (self::REF_ELEMENT_FUNCTION == $refElementType) {
-            $this->assertFalse(
-                function_exists($element),
-                "Function '$element', found in Reference ($min,$max), exists."
-            );
-
-        } elseif (self::REF_ELEMENT_CONSTANT == $refElementType) {
-            $this->assertFalse(
-                defined($element),
-                "Constant '$element', found in Reference ($min,$max), exists."
-            );
-
-        } elseif (self::REF_ELEMENT_CLASS == $refElementType) {
-            $this->assertFalse(
-                class_exists($element, false),
-                "Class '$element', found in Reference ($min,$max), exists."
-            );
-
-        } elseif (self::REF_ELEMENT_INTERFACE == $refElementType) {
-            $this->assertFalse(
-                interface_exists($element, false),
-                "Interface '$element', found in Reference ($min,$max), exists."
-            );
-
-        } elseif (self::REF_ELEMENT_METHOD == $refElementType) {
-            list ($object, $method) = explode('::', $element);
-            $this->assertFalse(
-                method_exists($object, $method),
-                "Class Method '$element', found in Reference, exists."
-            );
+        if (!empty($max) && $shouldBeThere) {
+            $shouldBeThere = version_compare(PHP_VERSION, $max, 'le');
         }
-    }
-
-    /**
-     * Generic Extension validator and producer
-     */
-    public function provideExtensionValues($methodName)
-    {
-        $this->checkExtension();
-
-        if ('testGetConstantsFromExtension' === $methodName) {
-            return $this->constantsFromExtensionProvider();
-
-        } elseif ('testGetClassConstantsFromExtension' === $methodName) {
-            return $this->classConstantsFromExtensionProvider();
-
-        } elseif ('testGetIniEntriesFromExtension' === $methodName) {
-            return $this->iniEntriesFromExtensionProvider();
-
-        } elseif ('testGetFunctionsFromExtension' == $methodName) {
-            return $this->functionsFromExtensionProvider();
-
-        } elseif ('testGetClassesFromExtension' === $methodName) {
-            return $this->classesFromExtensionProvider();
-
-        } elseif ('testGetInterfacesFromExtension' === $methodName) {
-            return $this->interfacesFromExtensionProvider();
-
-        } elseif ('testGetClassMethodsFromExtension' === $methodName) {
-            return $this->classMethodsFromExtensionProvider();
+        if (!empty($emin) && $shouldBeThere) {
+            $shouldBeThere = version_compare($extVersion, $emin, 'ge');
         }
+        if (!empty($emax) && $shouldBeThere) {
+            $shouldBeThere = version_compare($extVersion, $emax, 'le');
+        }
+
+        // Should be there except if set as optional
+        return $shouldBeThere;
     }
 
     /**
      * Provider to get INI entries from an extension
+     *
+     * @return Generator
      */
-    public function iniEntriesFromExtensionProvider()
+    private function iniEntriesFromExtensionProvider(): Generator
     {
-        $extension = new \ReflectionExtension(self::$ext);
+        $extension = $this->getReflectionExtension();
         $elements  = array_keys($extension->getINIEntries());
 
         foreach ($elements as $name) {
-            yield [$name];
+            yield $name;
         }
     }
 
     /**
      * Provider to get constants from an extension
+     *
+     * @return Generator
      */
-    public function constantsFromExtensionProvider()
+    private function constantsFromExtensionProvider(): Generator
     {
         $constants = get_defined_constants(true);
 
@@ -567,101 +324,113 @@ abstract class GenericTest extends \PHPUnit\Framework\TestCase
             array_pop($constants['Core']);
         }
 
-        $elements = isset($constants[self::$ext]) ? array_keys($constants[self::$ext]) : [];
+        $ext = self::$obj->getName();
+
+        $elements = isset($constants[$ext]) ? array_keys($constants[$ext]) : [];
 
         foreach ($elements as $name) {
-            yield [$name];
+            yield $name;
         }
     }
 
     /**
      * Provider to get functions from extension
+     *
+     * @return Generator
      */
-    public function functionsFromExtensionProvider()
+    private function functionsFromExtensionProvider(): Generator
     {
-        $elements = get_extension_funcs(strtolower(self::$ext));
+        $ext = self::$obj->getName();
+
+        $elements = get_extension_funcs(strtolower($ext));
         if (!is_array($elements)) {
             // can be NULL for ext without function
             $elements = [];
         }
 
         foreach ($elements as $name) {
-            yield [$name];
+            yield $name;
         }
     }
 
     /**
      * Provider to get classes from extension
+     *
+     * @return Generator
      */
-    public function classesFromExtensionProvider()
+    private function classesFromExtensionProvider(): Generator
     {
-        $extension = new \ReflectionExtension(self::$ext);
+        $extension = $this->getReflectionExtension();
         $classes   = array_unique($extension->getClassNames());
         $elements  = array_filter($classes, 'class_exists');
 
         foreach ($elements as $name) {
-            yield [$name];
+            yield $name;
         }
     }
 
     /**
      * Provider to get interfaces from extension
+     *
+     * @return Generator
      */
-    public function interfacesFromExtensionProvider()
+    private function interfacesFromExtensionProvider(): Generator
     {
-        $extension = new \ReflectionExtension(self::$ext);
+        $extension = $this->getReflectionExtension();
         $classes   = array_unique($extension->getClassNames());
         $elements  = array_filter($classes, 'interface_exists');
 
         foreach ($elements as $name) {
-            yield [$name];
+            yield $name;
         }
     }
 
     /**
      * Provider to get class constants from extension
+     *
+     * @return Generator
+     * @throws ReflectionException
      */
-    public function classConstantsFromExtensionProvider()
+    private function classConstantsFromExtensionProvider(): Generator
     {
-        $elements = [];
+        $extension = $this->getReflectionExtension();
+        $classes   = array_unique($extension->getClassNames());
+        $elements  = [];
 
-        if (in_array(self::$ext, self::$extensions)) {
-            $extension = new \ReflectionExtension(self::$ext);
-            $classes   = array_unique($extension->getClassNames());
-
-            foreach ($classes as $classname) {
-                $class = new \ReflectionClass($classname);
-                if ($class->getName() != $classname) {
-                    /* Skip class alias */
-                    continue;
-                }
-
-                $elements = $elements + array_map(
-                    function ($value) use ($classname) {
-                        return "$classname::$value";
-                    },
-                    array_keys($class->getConstants())
-                );
+        foreach ($classes as $classname) {
+            $class = new ReflectionClass($classname);
+            if ($class->getName() != $classname) {
+                /* Skip class alias */
+                continue;
             }
+
+            $elements = $elements + array_map(
+                function ($value) use ($classname) {
+                    return "$classname::$value";
+                },
+                array_keys($class->getConstants())
+            );
         }
 
         foreach ($elements as $name) {
-            yield [$name];
+            yield $name;
         }
     }
 
     /**
      * Provider to get class methods from extension
+     *
+     * @return Generator
+     * @throws ReflectionException
      */
-    public function classMethodsFromExtensionProvider()
+    private function classMethodsFromExtensionProvider(): Generator
     {
-        $extension = new \ReflectionExtension(self::$ext);
+        $extension = $this->getReflectionExtension();
         $classes   = array_unique($extension->getClassNames());
-
-        $elements = [];
+        $elements  = [];
 
         foreach ($classes as $classname) {
-            $class   = new \ReflectionClass($classname);
+            $class = new ReflectionClass($classname);
             if ($class->getName() != $classname) {
                 /* Skip class alias */
                 continue;
@@ -681,7 +450,8 @@ abstract class GenericTest extends \PHPUnit\Framework\TestCase
                     $method->getPrototype();
                     // don't check prototype methods
                     continue;
-                } catch (\ReflectionException $e) {
+                } catch (ReflectionException $e) {
+                    // none prototype for this method
                 }
 
                 $elements[] = $classname . '::' . $method->getName();
@@ -689,80 +459,77 @@ abstract class GenericTest extends \PHPUnit\Framework\TestCase
         }
 
         foreach ($elements as $name) {
-            yield [$name];
-        }
-    }
-
-    /**
-     * Check that a reference exists and initialize an instance
-     *
-     * @return void
-     */
-    public function checkExtension()
-    {
-        $parts = explode('\\', get_called_class());
-
-        self::$ext = $name = strtolower(
-            str_replace('ExtensionTest', '', end($parts))
-        );
-
-        // special case(s)
-        if ('zendopcache' === $name) {
-            self::$ext = $name = 'zend opcache';
-        }
-
-        if (!extension_loaded(self::$ext)) {
-            $this->markTestSkipped(
-                sprintf('Extension %s is required.', self::$ext)
-            );
+            yield $name;
         }
     }
 
     /**
      * Test than all referenced ini entries exists
      *
-     * @dataProvider provideReferenceValues
      * @group  reference
      * @return void
      */
-    public function testGetIniEntriesFromReference($name, $range, $refElementType)
+    public function testGetIniEntriesFromReference(): void
     {
-        $this->checkValuesFromReference(
-            $name,
-            $range,
-            $refElementType
-        );
+        foreach($this->provideReferenceValues(self::$obj->getIniEntries(), 'optionalcfgs') as $args) {
+            list($element, $range) = $args;
+
+            $shouldBeThere = $this->checkValuesFromReference(
+                $element,
+                $range,
+                self::$optionalcfgs,
+                self::$ignoredcfgs
+            );
+
+            if (null === $shouldBeThere) {
+                // test $element should be skipped because it was marked as optional or ignored
+                continue;
+            }
+
+            if ($shouldBeThere) {
+                $this->assertTrue(
+                    (false !== ini_get($element)),
+                    "INI '$element', found in Reference, does not exists."
+                );
+            } else {
+                $min = $range['php.min'];
+                $max = $range['php.max'];
+
+                $this->assertFalse(
+                    (false !== ini_get($element)),
+                    "INI '$element', found in Reference ($min, $max), exists."
+                );
+            }
+        }
     }
 
     /**
      * Test that each ini entries are defined in reference
      *
-     * @dataProvider provideExtensionValues
      * @group  reference
      * @return void
      */
-    public function testGetIniEntriesFromExtension($name)
+    public function testGetIniEntriesFromExtension(): void
     {
-        static $obj;
-        static $dict;
+        $ext = self::$obj->getName();
 
-        if ('internal' == self::$ext) {
+        if ('internal' === $ext) {
             // only Core is a valid extension name for API reflection
             return;
         }
 
-        if (!is_object($obj) || $obj->getName() !== $this->extension) {
-            $obj = new ExtensionFactory($this->extension);
-            $dict = $obj->getIniEntries();
-            $this->assertTrue(is_array($dict));
+        $generator = $this->iniEntriesFromExtensionProvider();
+        if (!$generator->valid()) {
+            return;
         }
+        $name = $generator->current();
 
         if (!in_array($name, self::$ignoredcfgs)) {
             $this->assertExtensionComponentHasKey(
                 $name,
-                $dict,
+                array_keys(self::$obj->getIniEntries()),
                 "Defined INI '$name' not known in Reference.",
-                $obj
+                self::$obj
             );
         }
     }
@@ -770,43 +537,86 @@ abstract class GenericTest extends \PHPUnit\Framework\TestCase
     /**
      * Test than all referenced functions exists
      *
-     * @dataProvider provideReferenceValues
      * @group  reference
      * @return void
      */
-    public function testGetFunctionsFromReference($name, $range, $refElementType)
+    public function testGetFunctionsFromReference(): void
     {
-        $this->checkValuesFromReference(
-            $name,
-            $range,
-            $refElementType
-        );
+        foreach ($this->provideReferenceValues(self::$obj->getFunctions(), 'optionalfunctions') as $args) {
+            list($element, $range) = $args;
+
+            $shouldBeThere = $this->checkValuesFromReference(
+                $element,
+                $range,
+                self::$optionalfunctions,
+                self::$ignoredfunctions
+            );
+
+            if (null === $shouldBeThere) {
+                // test $element should be skipped because it was marked as optional or ignored
+                continue;
+            }
+
+            if ($shouldBeThere) {
+                try {
+                    $function = new ReflectionFunction($element);
+                    $extensionName = $function->getExtensionName() ?: '';
+                    if (strcasecmp($extensionName, self::$obj->getName()) !== 0) {
+                        return;
+                    }
+                    $this->assertTrue(
+                        $function->isInternal(),
+                        "Function '$element', found in Reference, does not exists."
+                    );
+                } catch (ReflectionException $e) {
+                    // thrown if the given function does not exist.
+                    $this->assertTrue(
+                        false,
+                        "Function '$element', found in Reference, does not exists."
+                    );
+                }
+            } else {
+                $min = $range['php.min'];
+                $max = $range['php.max'];
+
+                try {
+                    $function = new ReflectionFunction($element);
+                    $extensionName = $function->getExtensionName() ?: '';
+                    if (strcasecmp($extensionName, self::$obj->getName()) !== 0) {
+                        return;
+                    }
+                    $this->assertFalse(
+                        $function->isInternal(),
+                        "Function '$element', found in Reference ($min, $max), exists."
+                    );
+                } catch (ReflectionException $e) {
+                    // thrown if the given function does not exist.
+                    return;
+                }
+            }
+        }
     }
 
     /**
      * Test that each functions are defined in reference
      *
-     * @dataProvider provideExtensionValues
      * @group  reference
      * @return void
      */
-    public function testGetFunctionsFromExtension($name)
+    public function testGetFunctionsFromExtension(): void
     {
-        static $obj;
-        static $dict;
-
-        if (!is_object($obj) || $obj->getName() !== $this->extension) {
-            $obj = new ExtensionFactory($this->extension);
-            $dict = $obj->getFunctions();
-            $this->assertTrue(is_array($dict));
+        $generator = $this->functionsFromExtensionProvider();
+        if (!$generator->valid()) {
+            return;
         }
+        $name = $generator->current();
 
         if (!in_array($name, self::$ignoredfunctions)) {
             $this->assertExtensionComponentHasKey(
                 $name,
-                $dict,
+                array_keys(self::$obj->getFunctions()),
                 "Defined function '$name' not known in Reference.",
-                $obj
+                self::$obj
             );
         }
     }
@@ -814,43 +624,63 @@ abstract class GenericTest extends \PHPUnit\Framework\TestCase
     /**
      * Test than all referenced constants exists
      *
-     * @dataProvider provideReferenceValues
      * @group  reference
      * @return void
      */
-    public function testGetConstantsFromReference($name, $range, $refElementType)
+    public function testGetConstantsFromReference(): void
     {
-        $this->checkValuesFromReference(
-            $name,
-            $range,
-            $refElementType
-        );
+        foreach ($this->provideReferenceValues(self::$obj->getConstants(), 'optionalconstants') as $args) {
+            list($element, $range) = $args;
+
+            $shouldBeThere = $this->checkValuesFromReference(
+                $element,
+                $range,
+                self::$optionalconstants,
+                self::$ignoredconstants
+            );
+
+            if (null === $shouldBeThere) {
+                // test $element should be skipped because it was marked as optional or ignored
+                continue;
+            }
+
+            if ($shouldBeThere) {
+                $this->assertTrue(
+                    defined($element),
+                    "Constant '$element', found in Reference, does not exists."
+                );
+            } else {
+                $min = $range['php.min'];
+                $max = $range['php.max'];
+
+                $this->assertFalse(
+                    defined($element),
+                    "Constant '$element', found in Reference ($min, $max), exists."
+                );
+            }
+        }
     }
 
     /**
      * Test that each constants are defined in reference
      *
-     * @dataProvider provideExtensionValues
      * @group  reference
      * @return void
      */
-    public function _testGetConstantsFromExtension($name)
+    public function testGetConstantsFromExtension(): void
     {
-        static $obj;
-        static $dict;
-
-        if (!is_object($obj) || $obj->getName() !== $this->extension) {
-            $obj = new ExtensionFactory($this->extension);
-            $dict = $obj->getConstants();
-            $this->assertTrue(is_array($dict));
+        $generator = $this->constantsFromExtensionProvider();
+        if (!$generator->valid()) {
+            return;
         }
+        $name = $generator->current();
 
-        // Test if each constants are in reference
         if (!in_array($name, self::$ignoredconstants)) {
-            $this->assertArrayHasKey(
+            $this->assertExtensionComponentHasKey(
                 $name,
-                $dict,
-                "Defined constant '$name' not known in Reference."
+                array_keys(self::$obj->getConstants()),
+                "Defined constant '$name' not known in Reference.",
+                self::$obj
             );
         }
     }
@@ -858,42 +688,70 @@ abstract class GenericTest extends \PHPUnit\Framework\TestCase
     /**
      * Test than all referenced classes exists
      *
-     * @dataProvider provideReferenceValues
      * @group  reference
      * @return void
      */
-    public function testGetClassesFromReference($name, $range, $refElementType)
+    public function testGetClassesFromReference(): void
     {
-        $this->checkValuesFromReference(
-            $name,
-            $range,
-            $refElementType
-        );
+        foreach ($this->provideReferenceValues(self::$obj->getClasses(), 'optionalclasses') as $args) {
+            list($element, $range) = $args;
+
+            $shouldBeThere = $this->checkValuesFromReference(
+                $element,
+                $range,
+                self::$optionalclasses,
+                self::$ignoredclasses
+            );
+
+            if (null === $shouldBeThere) {
+                // test $element should be skipped because it was marked as optional or ignored
+                continue;
+            }
+
+            if ($shouldBeThere) {
+                $this->assertTrue(
+                    class_exists($element, false),
+                    "Class '$element', found in Reference, does not exists."
+                );
+            } else {
+                $min = $range['php.min'];
+                $max = $range['php.max'];
+
+                try {
+                    $class = new ReflectionClass($element);
+                    $extensionName = $class->getExtensionName() ?: '';
+                    $this->assertFalse(
+                        (strcasecmp($extensionName, self::$obj->getName()) === 0),
+                        "Class '$element', found in Reference ($min, $max), exists."
+                    );
+                } catch (ReflectionException $e) {
+                    // thrown if the given class does not exist.
+                    return;
+                }
+            }
+        }
     }
 
     /**
      * Test that each classes are defined in reference
      *
-     * @dataProvider provideExtensionValues
      * @group  reference
      * @return void
      */
-    public function _testGetClassesFromExtension($name)
+    public function testGetClassesFromExtension(): void
     {
-        static $obj;
-        static $dict;
-
-        if (!is_object($obj) || $obj->getName() !== $this->extension) {
-            $obj = new ExtensionFactory($this->extension);
-            $dict = $obj->getClasses();
-            $this->assertTrue(is_array($dict));
+        $generator = $this->classesFromExtensionProvider();
+        if (!$generator->valid()) {
+            return;
         }
+        $name = $generator->current();
 
         if (!in_array($name, self::$ignoredclasses)) {
-            $this->assertArrayHasKey(
+            $this->assertExtensionComponentHasKey(
                 $name,
-                $dict,
-                "Defined class '$name' not known in Reference."
+                array_keys(self::$obj->getClasses()),
+                "Defined class '$name' not known in Reference.",
+                self::$obj
             );
         }
     }
@@ -901,83 +759,113 @@ abstract class GenericTest extends \PHPUnit\Framework\TestCase
     /**
      * Test than all referenced class methods exists
      *
-     * @dataProvider provideReferenceValues
      * @group  reference
      * @return void
      */
-    public function testGetClassMethodsFromReference($name, $range, $refElementType)
+    public function testGetMethodsFromReference(): void
     {
-        $this->checkValuesFromReference(
-            $name,
-            $range,
-            $refElementType
-        );
+        foreach ($this->provideReferenceValues(self::$obj->getMethods(), 'optionalmethods') as $args) {
+            list($element, $range) = $args;
+
+            $shouldBeThere = $this->checkValuesFromReference(
+                $element,
+                $range,
+                self::$optionalmethods,
+                self::$ignoredmethods
+            );
+
+            if (null === $shouldBeThere) {
+                // test $element should be skipped because it was marked as optional or ignored
+                continue;
+            }
+
+            try {
+                $method = new ReflectionMethod($element);
+                $extensionName = $method->getExtensionName() ?: '';
+                if (strcasecmp($extensionName, self::$obj->getName()) != 0) {
+                    return;
+                }
+                try {
+                    $method->getPrototype();
+                    // don't check prototype methods
+                    return;
+                } catch (ReflectionException $e) {
+                    // none prototype for this method
+                }
+                $methodExists = true;
+            } catch (ReflectionException $e) {
+                // thrown if the given method does not exist.
+                $methodExists = false;
+            }
+
+            if ($shouldBeThere) {
+                $this->assertTrue(
+                    $methodExists,
+                    "Class Method '$element', found in Reference, does not exists."
+                );
+            } else {
+                $this->assertFalse(
+                    $methodExists,
+                    "Class Method '$element', found in Reference, exists."
+                );
+            }
+        }
     }
 
     /**
      * Test that each class methods are defined in reference
      *
-     * @dataProvider provideExtensionValues
      * @group  reference
      * @return void
+     * @throws ReflectionException
      */
-    public function _testGetClassMethodsFromExtension($name)
+    public function testGetMethodsFromExtension(): void
     {
-        static $obj;
-        static $nonStaticMethods;
-        static $staticMethods;
-
-        if (!is_object($obj) || $obj->getName() !== $this->extension) {
-            $obj = new ExtensionFactory($this->extension);
-            $nonStaticMethods = $obj->getClassMethods();
-            $this->assertTrue(is_array($nonStaticMethods));
-            $staticMethods    = $obj->getClassStaticMethods();
-            $this->assertTrue(is_array($staticMethods));
+        $generator = $this->classMethodsFromExtensionProvider();
+        if (!$generator->valid()) {
+            return;
+        }
+        $methodName = $generator->current();
+        if (in_array($methodName, self::$ignoredmethods)) {
+            return;
         }
 
-        list ($classname, $name) = explode('::', $name);
+        $methods = self::$obj->getMethods();
 
-        if (isset($nonStaticMethods[$classname])) {
-            $this->assertArrayHasKey(
-                $name,
-                $nonStaticMethods[$classname],
-                "Defined method '$classname::$name' not known in Reference."
-            );
-        } else {
-            $this->assertArrayHasKey(
-                $name,
-                $staticMethods[$classname],
-                "Defined static method '$classname::$name' not known in Reference."
-            );
-        }
+        $this->assertNotEquals(
+            0,
+            count($methods),
+            'None method defined. Checks if `methods.json` file exists.'
+        );
+
+        $this->assertContains(
+            $methodName,
+            array_keys($methods),
+            "Defined method '$methodName' not known in Reference."
+        );
     }
 
     /**
      * Test that each class constants are defined in reference
      *
-     * @dataProvider provideExtensionValues
      * @group  reference
      * @return void
+     * @throws ReflectionException
      */
-    public function testGetClassConstantsFromExtension($name)
+    public function testGetClassConstantsFromExtension(): void
     {
-        static $obj;
-        static $dict;
-
-        if (!is_object($obj) || $obj->getName() !== $this->extension) {
-            $obj = new ExtensionFactory($this->extension);
-            $dict = $obj->getClassConstants();
-            $this->assertTrue(is_array($dict));
+        $generator = $this->classConstantsFromExtensionProvider();
+        if (!$generator->valid()) {
+            return;
         }
+        $constantName = $generator->current();
 
-        list ($classname, $name) = explode('::', $name);
-
-        if (!in_array($name, self::$ignoredconsts)) {
+        if (!in_array($constantName, self::$ignoredconsts)) {
             $this->assertExtensionComponentHasKey(
-                $name,
-                $dict[$classname],
-                "Defined class constant '$classname::$name' not known in Reference.",
-                $obj
+                $constantName,
+                array_keys(self::$obj->getClassConstants()),
+                "Defined class constant '$constantName' not known in Reference.",
+                self::$obj
             );
         }
     }
@@ -985,51 +873,81 @@ abstract class GenericTest extends \PHPUnit\Framework\TestCase
     /**
      * Test than all referenced interfaces exists
      *
-     * @dataProvider provideReferenceValues
      * @group  reference
      * @return void
      */
-    public function testGetInterfacesFromReference($name, $range, $refElementType)
+    public function testGetInterfacesFromReference(): void
     {
-        $this->checkValuesFromReference(
-            $name,
-            $range,
-            $refElementType
-        );
+        foreach ($this->provideReferenceValues(self::$obj->getInterfaces(), 'optionalinterfaces') as $args) {
+            list($element, $range) = $args;
+
+            $shouldBeThere = $this->checkValuesFromReference(
+                $element,
+                $range,
+                self::$optionalinterfaces,
+                self::$ignoredinterfaces
+            );
+
+            if (null === $shouldBeThere) {
+                // test $element should be skipped because it was marked as optional or ignored
+                continue;
+            }
+
+            if ($shouldBeThere) {
+                $this->assertTrue(
+                    interface_exists($element, false),
+                    "Interface '$element', found in Reference, does not exists."
+                );
+            } else {
+                $min = $range['php.min'];
+                $max = $range['php.max'];
+
+                try {
+                    $class = new ReflectionClass($element);
+                    if (!$class->isInterface()) {
+                        return;
+                    }
+                    $extensionName = $class->getExtensionName() ?: '';
+                    $this->assertFalse(
+                        (strcasecmp($extensionName, self::$obj->getName()) === 0),
+                        "Interface '$element', found in Reference ($min, $max), exists."
+                    );
+                } catch (ReflectionException $e) {
+                    // thrown if the given interface does not exist.
+                    return;
+                }
+            }
+        }
     }
 
     /**
      * Test that each interface is defined in reference
      *
-     * @dataProvider provideExtensionValues
      * @group  reference
      * @return void
      */
-    public function testGetInterfacesFromExtension($name)
+    public function testGetInterfacesFromExtension(): void
     {
-        static $obj;
-        static $dict;
-
-        if (!is_object($obj) || $obj->getName() !== $this->extension) {
-            $obj = new ExtensionFactory($this->extension);
-            $dict = $obj->getInterfaces();
-            $this->assertTrue(is_array($dict));
+        $generator = $this->interfacesFromExtensionProvider();
+        if (!$generator->valid()) {
+            return;
         }
+        $name = $generator->current();
 
         if (!in_array($name, self::$ignoredinterfaces)) {
             $this->assertExtensionComponentHasKey(
                 $name,
-                $dict,
+                array_keys(self::$obj->getInterfaces()),
                 "Defined interface '$name' not known in Reference.",
-                $obj
+                self::$obj
             );
         }
     }
 
-    public function assertExtensionComponentHasKey($key, $array, $message, $obj)
+    private function assertExtensionComponentHasKey($needle, $haystack, $message, $obj): void
     {
         try {
-            $this->assertArrayHasKey($key, $array, $message);
+            $this->assertContains($needle, $haystack, $message);
         } catch (ExpectationFailedException $e) {
             $warning = $this->checkUpdateExtension($obj);
 
@@ -1041,28 +959,35 @@ abstract class GenericTest extends \PHPUnit\Framework\TestCase
         }
     }
 
-    protected function checkUpdateExtension($obj)
+    private function checkUpdateExtension(Extension $obj): ?string
     {
-        $currentVersion = $obj->getCurrentVersion();
-
+        $currentVersion = $this->getExtensionVersion($obj->getName());
         if ($currentVersion === false) {
             // extension did not provide any version information
-            return;
+            return null;
         }
 
-        $releases = array_keys($obj->getReleases());
-
-        $latestReleaseReferenced = array_pop($releases);
+        $latestReleaseReferenced = $obj->getLastRelease()->getVersion();
         // check if extension installed is more recent than the one declared in compatinfo-db
         if (version_compare($currentVersion, $latestReleaseReferenced, 'le')) {
-            return;
+            return null;
         }
 
         return sprintf(
             'Extension %s tested is version %s, while latest version referenced is %s and has need update.',
             $obj->getName(),
-            $obj->getCurrentVersion(),
+            $currentVersion,
             $latestReleaseReferenced
         );
+    }
+
+    private function getReflectionExtension(): ReflectionExtension
+    {
+        $name = self::$obj->getName();
+        // special case(s)
+        if ('opcache' === $name) {
+            $name = 'zend opcache';
+        }
+        return new ReflectionExtension($name);
     }
 }
