@@ -1,9 +1,22 @@
-<?php
+<?php declare(strict_types=1);
 
-declare(strict_types=1);
+/**
+ * Marker interface for command handlers.
+ *
+ * PHP version 7
+ *
+ * @category   PHP
+ * @package    PHP_CompatInfo_Db
+ * @author     Laurent Laville <pear@laurent-laville.org>
+ * @license    https://opensource.org/licenses/BSD-3-Clause The 3-Clause BSD License
+ * @link       http://bartlett.laurent-laville.org/php-compatinfo/
+ */
 
 namespace Bartlett\CompatInfoDb\Application\Command;
 
+/**
+ * @since 3.0.0
+ */
 interface CommandHandlerInterface
 {
 }
