@@ -1,44 +1,57 @@
-<?php
+<?php declare(strict_types=1);
 
-declare(strict_types=1);
+/**
+ * Symfony Console Application to handle the CompatInfo database.
+ *
+ * PHP version 7
+ *
+ * @category   PHP
+ * @package    PHP_CompatInfo_Db
+ * @author     Laurent Laville <pear@laurent-laville.org>
+ * @license    https://opensource.org/licenses/BSD-3-Clause The 3-Clause BSD License
+ * @link       http://bartlett.laurent-laville.org/php-compatinfo/
+ */
 
 namespace Bartlett\CompatInfoDb\Presentation\Console;
 
-use Bartlett\CompatInfoDb\DatabaseFactory;
-use Bartlett\CompatInfoDb\Presentation\Console\Command\BuildExtensionCommand;
-use Bartlett\CompatInfoDb\Presentation\Console\Command\DiagnoseCommand;
-use Bartlett\CompatInfoDb\Presentation\Console\Command\InitCommand;
-use Bartlett\CompatInfoDb\Presentation\Console\Command\ListCommand;
-use Bartlett\CompatInfoDb\Presentation\Console\Command\PublishCommand;
-use Bartlett\CompatInfoDb\Presentation\Console\Command\ReleaseCommand;
-use Bartlett\CompatInfoDb\Presentation\Console\Command\ShowCommand;
-
 use PackageVersions\Versions;
 
-use Psr\Container\ContainerInterface;
-
-use Symfony\Component\Console\CommandLoader\CommandLoaderInterface;
-use Symfony\Component\Console\CommandLoader\ContainerCommandLoader;
+use Symfony\Component\Config\Exception\FileLocatorFileNotFoundException;
+use Symfony\Component\Config\FileLocator;
+use Symfony\Component\Console\Application as SymfonyApplication;
+use Symfony\Component\Console\Input\ArgvInput;
+use Symfony\Component\Console\Input\InputDefinition;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Output\ConsoleOutput;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+use Symfony\Component\DependencyInjection\Loader\PhpFileLoader;
 
-use PDO;
+use Phar;
+use function basename;
+use function dirname;
+use function explode;
+use function is_callable;
+use function sprintf;
+use function strpos;
+use function substr_count;
 
 /**
- * Symfony Console Application to handle the SQLite compatinfo database.
+ * @since Release 2.0.0RC1
  */
-class Application extends \Symfony\Component\Console\Application
+class Application extends SymfonyApplication implements ApplicationInterface
 {
-    public const NAME = 'Database handler for CompatInfo';
-    public const VERSION = '2.19.0';
-
-    /** @var string */
-    private $baseDir;
-
     /** @var ContainerInterface  */
     private $container;
 
-    public function __construct(ContainerInterface $container, string $version = 'UNKNOWN')
+    /**
+     * Application constructor.
+     *
+     * @param string $version (optional) auto-detect
+     */
+    public function __construct(string $version = 'UNKNOWN')
     {
         if ('UNKNOWN' === $version) {
             // composer or git outside world strategy
@@ -55,79 +68,134 @@ class Application extends \Symfony\Component\Console\Application
             }
         }
         parent::__construct(self::NAME, $version);
+    }
 
+    /**
+     * {@inheritDoc}
+     */
+    protected function getDefaultInputDefinition(): InputDefinition
+    {
+        $definition = parent::getDefaultInputDefinition();
+        if (Phar::running()) {
+            $definition->addOption(
+                new InputOption(
+                    'manifest',
+                    null,
+                    InputOption::VALUE_NONE,
+                    'Show which versions of dependencies are bundled'
+                )
+            );
+            // handle external configuration files is not allowed with PHAR distribution
+            return $definition;
+        }
+        $definition->addOption(
+            new InputOption(
+                'config',
+                'c',
+                InputOption::VALUE_REQUIRED,
+                'Read configuration from PHP file'
+            )
+        );
+        $definition->addOption(
+            new InputOption(
+                'no-configuration',
+                null,
+                InputOption::VALUE_NONE,
+                'Ignore current configuration and run with only required services (config/set/common.php)'
+            )
+        );
+        $definition->addOption(
+            new InputOption(
+                'profile',
+                null,
+                InputOption::VALUE_NONE,
+                'Display timing and memory usage information'
+            )
+        );
+        return $definition;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function run(InputInterface $input = null, OutputInterface $output = null)
+    {
+        if (null === $input) {
+            if ($this->container->has(InputInterface::class)) {
+                $input = $this->container->get(InputInterface::class);
+            } else {
+                $input = new ArgvInput();
+            }
+
+
+            if ($input->hasParameterOption('--no-configuration')) {
+                $configFile = 'config/set/common.php';
+            } else {
+                $configFile = $input->getParameterOption('-c');
+            }
+
+            if (false === $configFile) {
+                $configFile = $input->getParameterOption('--config');
+            }
+            if (false !== $configFile) {
+                $containerBuilder = new ContainerBuilder();
+                try {
+                    $loader = new PhpFileLoader($containerBuilder, new FileLocator(dirname($configFile)));
+                    $loader->load(basename($configFile));
+                } catch (FileLocatorFileNotFoundException $e) {
+                    $output = new ConsoleOutput();
+                    $this->renderThrowable($e, $output);
+                    return 1;
+                }
+                $containerBuilder->compile();
+                $this->setContainer($containerBuilder);
+            }
+        }
+
+        if (null === $output) {
+            if ($this->container->has(OutputInterface::class)) {
+                $output = $this->container->get(OutputInterface::class);
+            } else {
+                $output = new ConsoleOutput();
+            }
+        }
+
+        if ($input->hasParameterOption('--manifest')) {
+            $phar = new Phar('compatinfo-db.phar');
+            $manifest = $phar->getMetadata();
+            if (is_callable($manifest)) {
+                $manifest = $manifest();
+            }
+            $output->writeln($manifest);
+            return 0;
+        }
+
+        return parent::run($input, $output);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function setContainer(ContainerInterface $container = null)
+    {
         $this->container = $container;
-        $this->setCommandLoader($this->createCommandLoader($container));
-        $this->baseDir = dirname(__DIR__, 3);
     }
 
-    public function getDbFilename() : string
-    {
-        return $this->baseDir . DIRECTORY_SEPARATOR . 'data' . DIRECTORY_SEPARATOR . 'compatinfo.sqlite';
-    }
-
-    public function getRefDir() : string
-    {
-        return $this->baseDir . DIRECTORY_SEPARATOR . 'data' . DIRECTORY_SEPARATOR . 'references';
-    }
-
-    public function getLongVersion() : string
+    /**
+     * {@inheritDoc}
+     */
+    public function getLongVersion(): string
     {
         if ('UNKNOWN' !== $this->getName()) {
             if ('UNKNOWN' !== $this->getVersion()) {
-                $v = $this->getDbVersions();
-
                 return sprintf(
-                    '<info>%s</info> version <comment>%s</comment> DB built <comment>%s</comment>',
+                    '<info>%s</info> version <comment>%s</comment>',
                     $this->getName(),
-                    $this->getVersion(),
-                    $v['build.string']
+                    $this->getVersion()
                 );
             }
-
             return $this->getName();
         }
-
         return 'Console Tool';
-    }
-
-    public function getDbVersions() : array
-    {
-        $pdo = DatabaseFactory::create('sqlite');
-
-        $stmt = $pdo->prepare(
-            'SELECT build_string as "build.string", build_date as "build.date", build_version as "build.version"' .
-            ' FROM bartlett_compatinfo_versions'
-        );
-        $stmt->execute();
-        return $stmt->fetch(PDO::FETCH_ASSOC);
-    }
-
-    public function doRun(InputInterface $input, OutputInterface $output)
-    {
-        $this->container->set(InputInterface::class, $input);
-        $this->container->set(OutputInterface::class, $output);
-
-        return parent::doRun($input, $output);
-    }
-    /**
-     * @param ContainerInterface $container
-     * @return CommandLoaderInterface
-     * @see https://symfony.com/doc/current/console/lazy_commands.html#containercommandloader
-     */
-    private function createCommandLoader(ContainerInterface $container): CommandLoaderInterface
-    {
-        return new ContainerCommandLoader(
-            $container,
-            [
-                BuildExtensionCommand::NAME => BuildExtensionCommand::class,
-                DiagnoseCommand::NAME => DiagnoseCommand::class,
-                InitCommand::NAME => InitCommand::class,
-                ListCommand::NAME => ListCommand::class,
-                PublishCommand::NAME => PublishCommand::class,
-                ReleaseCommand::NAME => ReleaseCommand::class,
-                ShowCommand::NAME => ShowCommand::class,
-            ]
-        );
     }
 }
