@@ -1,79 +1,39 @@
-<?php
+<?php declare(strict_types=1);
 
-declare(strict_types=1);
+/**
+ * Base class for all application commands.
+ *
+ * PHP version 7
+ *
+ * @category   PHP
+ * @package    PHP_CompatInfo_Db
+ * @author     Laurent Laville <pear@laurent-laville.org>
+ * @license    https://opensource.org/licenses/BSD-3-Clause The 3-Clause BSD License
+ * @link       http://bartlett.laurent-laville.org/php-compatinfo/
+ */
 
 namespace Bartlett\CompatInfoDb\Presentation\Console\Command;
 
-use League\Tactician\CommandBus;
-use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Helper\Table;
+use Bartlett\CompatInfoDb\Application\Command\CommandBusInterface;
+use Bartlett\CompatInfoDb\Application\Query\QueryBusInterface;
 
-class AbstractCommand extends Command
+use Symfony\Component\Console\Command\Command;
+
+/**
+ * @since Release 2.0.0RC1
+ */
+abstract class AbstractCommand extends Command
 {
-    /** @var CommandBus */
+    /** @var CommandBusInterface */
     protected $commandBus;
 
-    public function __construct(CommandBus $commandBus)
+    /** @var QueryBusInterface */
+    protected $queryBus;
+
+    public function __construct(CommandBusInterface $commandBus, QueryBusInterface $queryBus)
     {
         parent::__construct();
         $this->commandBus = $commandBus;
-    }
-
-    /**
-     * Prints the database current build version
-     *
-     * @param OutputInterface $output Console Output concrete instance
-     *
-     * @return void
-     */
-    protected function printDbBuildVersion(OutputInterface $output) : void
-    {
-        $output->writeln(
-            sprintf(
-                '<info>Reference Database Version</info> => %s%s',
-                $this->getApplication()->getDbVersions()['build.version'],
-                PHP_EOL
-
-            )
-        );
-    }
-
-    /**
-     * Helper that convert analyser results to a console table
-     *
-     * @param OutputInterface $output  Console Output concrete instance
-     * @param array           $headers All table headers
-     * @param array           $rows    All table rows
-     * @param string          $style   The default style name to render tables
-     *
-     * @return void
-     */
-    protected function tableHelper(OutputInterface $output, array $headers, array $rows, string $style = 'compact') : void
-    {
-        $table = new Table($output);
-        $table->setStyle($style)
-            ->setHeaders($headers)
-            ->setRows($rows)
-            ->render()
-        ;
-    }
-
-    /**
-     * Helper that convert an array key-value pairs to a console report.
-     *
-     * See Structure and Loc analysers for implementation examples
-     *
-     * @param OutputInterface $output Console Output concrete instance
-     * @param array           $lines  Any analyser formatted metrics
-     *
-     * @return void
-     */
-    protected function printFormattedLines(OutputInterface $output, array $lines) : void
-    {
-        foreach ($lines as $ident => $contents) {
-            list ($format, $args) = $contents;
-            $output->writeln(vsprintf($format, $args));
-        }
+        $this->queryBus = $queryBus;
     }
 }
